+++ Assets/Scripts/DURVAL/NoteView.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public struct NoteView
 {
     public GameObject GameObject;
-    public float X;
+    public double X;
     public Pitch Pitch;
-    public float beatNumber;
-    public float noteTimeInSeconds;
+    public double beatNumber;
+    public double noteTimeInSeconds;
     public bool lastNote;
     public bool isRest;
 }
